==> processing/01-proc-data.qmd <==
---
title: "Data preparation"
subtitle: "Changes in Beliefs about Meritocracy and Preferences for Market Justice in the Chilean School Context"
author: EDUMER Team
date: today
lang: en
fontsize: 12pt
format:
  html:
    toc: true
    toc-location: right
    toc-depth: 2
    toc-expand: 2
    toc-title: Contents
    number-sections: true
    number-depth: 3
    theme:
      - cosmo
      - "edumer.scss"
    code-link: true
    title-block-banner: true
  pdf:
    number-sections: true
    number-depth: 3
editor_options: 
  chunk_output_type: console
---

# Presentation

This is the data preparation code for the paper "Changes in Beliefs about Meritocracy and Preferences for Market Justice in the Chilean School Context". The prepared data is `edumer_students_long.RData`.

```{r}
#| label: set
#| echo: false
#| message: false
#| warning: false

library(knitr)
knitr::opts_chunk$set(echo = TRUE, include = TRUE, warning = FALSE, message = FALSE)

table_format <- if(is_html_output()) {
  "html"
} else if(is_latex_output()) {
  "latex"
}
table_format2 <- if(is_html_output()) {
  T
} else if(is_latex_output()) {
  F
}

options(kableExtra.html.bsTable = T)
options(knitr.kable.NA = "")
```

# Libraries

```{r}
#| label: libraries

if (! require("pacman")) install.packages("pacman")

pacman::p_load(tidyverse,
               sjmisc, 
               here,
               sjlabelled,
               naniar,
               car,
               kableExtra,
               summarytools)

options(scipen=999)
rm(list = ls())

```

# Data

```{r}
#| label: data

load(url("https://github.com/educacion-meritocracia/edumer-data/raw/main/output/data/edumer_students_long.RData"))
load(url("https://github.com/educacion-meritocracia/edumer-data/raw/main/output/data/db_proc_parents.RData"))

db_long <- edumer_students_long

rm(edumer_students_long)

glimpse(db_long)
```

# Processing

## Select

```{r}
#| label: select

db_long <- db_long %>% 
  select(id_estudiante,
         ola,
         consent = consentimiento,
         curse_level = nivel_estudiante,
         perc_effort = p1_1,
         perc_talent = p1_2,
         perc_rich_parents = p1_3,
         perc_contact = p1_4,
         pref_effort = p1_5,
         pref_talent = p1_6,
         pref_rich_parents = p1_7,
         pref_contact = p1_8,
         school_effort = p2_1,
         school_talent = p2_2,
         inequality = p9_1,
         redistribucion = p9_2,
         just_educ = p9_3,
         just_health = p9_4,
         just_pension = p9_5,
         gender = p20,
         age = p21_ano,
         educ_m = p26,
         educ_p = p27,
         books = p30)
```

# Filter

```{r}
#| label: filter

db_long <- db_long %>% 
  filter(consent == 1) %>% select(-consent)
```

# Recode and transform

```{r}
#| label: recode and transform 

# level
frq(db_long$curse_level)

db_long <- db_long %>% 
  mutate(
    cohort_level = if_else(curse_level %in% c("6to", "7mo"), "Primary", "Secondary"),
    cohort_level = factor(cohort_level, 
                          labels = c("Primary", "Secondary")),
    curse_level = factor(curse_level, 
                         levels = c("6to", "7mo", "1ro", "2do"))
  )

# merit-scale society
frq(db_long$perc_effort)
frq(db_long$perc_talent)
frq(db_long$perc_rich_parents)
frq(db_long$perc_contact)
frq(db_long$pref_effort)
frq(db_long$pref_talent)
frq(db_long$pref_rich_parents)
frq(db_long$pref_contact)


labels1 <- c("Strongly desagree" = 1, 
             "Disagree" = 2, 
             "Agree" = 3, 
             "Strongly agree" = 4, 
             "Dont know" = 88, 
             "No response" = 99)

db_long <- db_long %>% 
  mutate_at(.vars = (4:11),.funs = ~ sjlabelled::set_labels(., labels = labels1))

db_long <- db_long %>% 
  mutate_at(.vars = (4:11),.funs = ~ set_na(., na = c(88,99)))

# merit school
frq(db_long$school_effort)
frq(db_long$school_talent)

db_long <- db_long %>% 
  mutate_at(.vars = (12:13),.funs = ~ sjlabelled::set_labels(., labels = labels1))

db_long <- db_long %>% 
  mutate_at(.vars = (12:13),.funs = ~ set_na(., na = c(88,99)))


# percep inequality
frq(db_long$inequality)

db_long <- db_long %>% 
  mutate(inequality = sjlabelled::set_labels(inequality, labels = labels1),
         inequality = set_na(inequality, na = c(88,99)))

# market justice preference
frq(db_long$just_educ)
frq(db_long$just_health)
frq(db_long$just_pension)

db_long <- db_long %>% 
  mutate_at(.vars = (16:18),.funs = ~ sjlabelled::set_labels(., labels = labels1))

db_long <- db_long %>% 
  mutate_at(.vars = (16:18),.funs = ~ set_na(., na = c(88,99)))

db_long$mjp <- rowMeans(x = db_long[16:18], na.rm = T)
db_long$mjp <- if_else(is.nan(db_long$mjp), NA, db_long$mjp)

summary(db_long$mjp)

# gender
frq(db_long$gender)

db_long$gender <- car::recode(db_long$gender, 
                               recodes = c("'H'= 'Male'; 'M' = 'Female'; 'O' = 'Other'"))

db_long$gender <- factor(db_long$gender, 
                               levels = c("Male", "Female", "Other"))

# age
frq(db_long$age)

db_long$age <- 2024 - db_long$age

db_long$age2 <- db_long$age^2

# ses
frq(db_long$books)

db_long <- db_long %>% 
  mutate(books = case_when(books <= 2 ~ "Less than 25",
                           books >= 3 & books <= 6 ~ "More than 25",
                           TRUE ~ NA_character_),
         books = factor(books, levels = c("Less than 25", "More than 25")))


db_long <- left_join(
  db_long,
  db_parents[,c("id_estudiante_apoderado", "id_apoderado", "p12_1_apoderado")], 
  by = c("id_estudiante" = "id_estudiante_apoderado")

)

frq(db_long$p12_1_apoderado)
frq(db_long$educ_m)
frq(db_long$educ_p)

db_long <- db_long %>% 
  mutate(
    across(
      .cols = c(educ_m, educ_p, p12_1_apoderado),
      .fns = ~ set_na(., na = c(88,99))
    )
  )

db_long <- db_long %>%
  rowwise() %>%
  mutate(
    aux1 = case_when(
      !is.na(educ_p) & !is.na(educ_m) ~ max(educ_p, educ_m),  
      is.na(educ_p)   & !is.na(educ_m) ~ educ_m,               
      !is.na(educ_p)  &  is.na(educ_m) ~ educ_p,               
      TRUE                                  ~ NA_real_         
    )
  ) %>% 
  ungroup()

frq(db_long$aux1)
frq(db_long$p12_1_apoderado)

db_long <- db_long %>% 
  group_by(id_estudiante) %>% 
  rowwise() %>% 
  mutate(parental_educ = if_else(is.na(p12_1_apoderado), aux1, p12_1_apoderado)) %>% 
  ungroup()

frq(db_long$parental_educ)

db_long <- db_long %>% 
  rowwise() %>% 
  mutate(parental_educ = case_when(parental_educ == 1 ~ "8th grade or less",
                              parental_educ == 2 ~ "8th grade or less",
                              parental_educ == 3 ~ "8th grade or less",
                              parental_educ == 4 ~ "Secondary education",
                              parental_educ == 5 ~ "Secondary education",
                              parental_educ == 6 ~ "Technical higher education",
                              parental_educ == 7 ~ "Technical higher education",
                              parental_educ == 8 ~ "Universitary or posgraduate",
                              parental_educ == 9 ~ "Universitary or posgraduate",
                              parental_educ == 10 ~ "Universitary or posgraduate",
                              is.na(parental_educ) ~ "Missing")) %>% 
  ungroup()

db_long$parental_educ <- factor(db_long$parental_educ, levels = c("8th grade or less", "Secondary education", "Technical higher education", "Universitary or posgraduate", "Missing"))


```


```{r}
#| label: etiquetas

db_long$curse_level <- set_label(db_long$curse_level,label = "Student level")
db_long$cohort_level <- set_label(db_long$cohort_level,label = "Student cohort")
db_long$gender <- set_label(db_long$gender,label = "Gender")
db_long$age <- set_label(db_long$age,label = "Age")
db_long$age2 <- set_label(db_long$age2,label = "Age^2")
db_long$books <- set_label(db_long$books,label = "Books at home")
db_long$mjp <- set_label(db_long$mjp,label = "Market Justice Preferences")
db_long$parental_educ <- set_label(db_long$parental_educ,label = "Parental educational level")
db_long$id_estudiante <- set_label(db_long$id_estudiante,label = "Unique student identifier")
db_long$ola <- set_label(db_long$ola,label = "Wave")

db_long$perc_effort <- set_label(db_long$perc_effort,label = "In Chile, people are rewarded for their efforts")
db_long$perc_talent <- set_label(db_long$perc_talent,label = "In Chile, people are rewarded for their intelligence and ability")
db_long$perc_rich_parents <- set_label(db_long$perc_rich_parents,label = "In Chile, those with wealthy parents fare much better in life")
db_long$perc_contact <- set_label(db_long$perc_contact,label = "In Chile, those with good contacts do better in life")

db_long$pref_effort <- set_label(db_long$pref_effort,label = "Those who put in more effort should receive greater rewards than those who put in less effort")
db_long$pref_talent <- set_label(db_long$pref_talent,label = "Those who have more talent should receive greater rewards than those who have less talent")
db_long$pref_rich_parents <- set_label(db_long$pref_rich_parents,label = "It’s fine that those whose parents are wealthy do well in life")
db_long$pref_contact <- set_label(db_long$pref_contact,label = "It’s fine that those who have good connections do well in life")

db_long$school_effort <- set_label(db_long$school_effort,label = "In this school, those who put in effort get good grades")
db_long$school_talent <- set_label(db_long$school_talent,label = "In this school, those who are intelligent get good grades")
db_long$school_merit <- set_label(db_long$school_merit,label = "In this school, students get the grades that they deserve")

db_long$just_educ <- set_label(db_long$just_educ,label = "It’s good that those who can pay more receive a better education")
db_long$just_health <- set_label(db_long$just_health,label = "It’s good that those who can pay more have better access to healthcare")
db_long$just_pension <- set_label(db_long$just_pension,label = "It’s good that in Chile, people with higher incomes can have better pensions than those with lower incomes")

db_long$inequality <- set_label(db_long$inequality,label = "Income diferences between rich and poor in Chile is too large")
db_long$redistribucion <- set_label(db_long$redistribucion,label = "It is the government's responsibility to reduce the economic gap between high-income people and low-income people")

  

db_proc <- db_long %>% 
  select(-c(educ_m, educ_p, aux1, p12_1_apoderado, id_apoderado)) %>% 
  rename(id_student = id_estudiante,
         wave = ola)

```


# Missings

```{r}
#| label: missings 

colSums(is.na(db_proc))

n_miss(db_proc)

prop_miss(db_proc)*100

miss_var_summary(db_proc)

vis_miss(db_proc) + theme(axis.text.x = element_text(angle=80))

db_proc <- na.omit(db_proc)

db_proc <- db_proc %>%
  group_by(id_student) %>%
  filter(n() > 1) %>%
  ungroup()
```


```{r}
print(summarytools::dfSummary(db_proc), method="render")
```


# Save and export

```{r}
save(db_proc, file = here("output/data/db_proc.RData"))
```
